requirements.txt
azure-functions
requests
pandas
numpy
azure-storage-blob
azure-identity
azure-core